Scope task generator to maintenance_tasks
