+++ lib/generators/maintenance_tasks/templates/task.rb.tt
