+++ lib/generators/maintenance_tasks/templates/task_test.rb.tt
